requirements.txt
pytest==8.0.0
python-dotenv==1.0.1
Requests==2.31.0
